// src/lib/api/models/ContentDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ContentDto = {
    id?: number;
    title?: string | null;
    content?: string | null;
    image?: string | null;
    slug?: string | null;
    createdOn?: string;
    lastEdited?: string;
};